cli/describe_image_cli.py
import argparse
import mimetypes
from google import genai
from google.genai import types
from dotenv import load_dotenv
import os

def main():
    desc_img_parser = argparse.ArgumentParser(description="Describe Image CLI")
    desc_img_parser.add_argument("--image", type=str, help="Path to image")
    desc_img_parser.add_argument("--query", type=str, help="Query")
    args = desc_img_parser.parse_args()
    
    mime, _ = mimetypes.guess_type(args.image)
    mime = mime or "image/jpeg"
    with open(args.image, "rb") as img:
        image_bytes = img.read()
        
    load_dotenv()
    api_key = os.environ.get("GEMINI_API_KEY")
    client = genai.Client(api_key=api_key)
    parts = [
        get_prompt(),
        types.Part.from_bytes(data=image_bytes, mime_type=mime),
        args.query.strip(),
    ]

    response = client.models.generate_content(model="gemini-2.5-flash", contents=parts)
    print(f"Rewritten query: {response.text.strip()}")
    if response.usage_metadata is not None:
        print(f"Total tokens:    {response.usage_metadata.total_token_count}")

def get_prompt():
    return f"""
        Given the included image and text query, rewrite the text query to improve search results from a movie database. Make sure to:
        - Synthesize visual and textual information
        - Focus on movie-specific details (actors, scenes, style, etc.)
        - Return only the rewritten query, without any additional commentary
    """

if __name__ == "__main__":
    main()